Refact on some class names

+++ NodeExtension.cs
@@ -4,12 +4,12 @@ using System.Linq;
 
 public static class NodeExtension
 {
-    public static void EditText(this Node n, string path, Func<int, int> cb)
+    public static void EditText(this Node n, string path, Func<float, float> cb)
     {
         if (!n.GetTree().Root.GetChild(0).HasNode(path)) return;
         var textEditor = n.GetParent().GetNode<TextEdit>(path);
         var texts = textEditor.Text.Split(" ");
-        if (int.TryParse(texts[^1], out int val))
+        if (float.TryParse(texts[^1], out float val))
         {
             texts[^1] = (cb(val)).ToString();
         }
